ASDK-1397: Add code to the storyblock app
