+++ storyblock/app/src/main/res/values/strings.xml
@@ -1,8 +1,6 @@
 <resources>
-    <string name="app_name">FW Discovery</string>
-
-    <string name="source_type">Source Type:</string>
-    <string name="channel_id">Channel ID:</string>
-    <string name="playlist_id">Playlist ID:</string>
-    <string name="categories">Categories:</string>
+    <string name="app_name">Storyblock</string>
+    <string name="lorem_ipsum">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum</string>
+    <string name="play">Play</string>
+    <string name="pause">Pause</string>
 </resources>
