+++ storyblock/app/src/main/res/values/themes.xml
@@ -1,6 +1,6 @@
 <resources>
     <!-- Base application theme. -->
-    <style name="Theme.FireworkExample" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
+    <style name="Theme.FireworkExample" parent="Theme.MaterialComponents.DayNight.NoActionBar">
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/purple_500</item>
         <item name="colorPrimaryVariant">@color/purple_700</item>
